Cleanup: Confusing and varied naming of 'last_control_point_index'
Function curves::nurbs::calculate_basis_cache() varies the name of the
variable 'last_control_point_index '. Name is also confusing and not
descriptive as it  refers to the total number of points including
padded/repeated points for cyclic curves (repeated by wrapping the
index by the size of the array).

Use of 'points_num' generally also refers to the size of the array and
not total number of wrapped points.

Pull Request: https://projects.blender.org/blender/blender/pulls/143921