Fix #141173: Inpaint node inconsistent on GPU
The Inpaint node produces inconsistent results across different GPUs,
where it produces different results on Intel Windows GPU platforms. This
is due to a mismatch between the precision declared for the shader and
that of the allocated textures. To fix this, we make sure the shader
uses the same precision as the allocates results.

Pull Request: https://projects.blender.org/blender/blender/pulls/143793