Fix #143839: missing warning when connecting field to single value
The behavior of the field inferencing changed slightly. So instead of checking
if the link connects a socket that has a field to one that does not, check the
source is a field and the target does not support one. Note the difference
between "having a field" and "supporting a field".

This was probably broken in 7b7c63001858280ffec1a13c2751c22cbcf27686.

Pull Request: https://projects.blender.org/blender/blender/pulls/143975