Cleanup: VSE: Remove `StripData.start/endstill`
Not to be confused with `Strip.start/endstill` which were renamed with
a legacy suffix. The `StripData` version here isn't present anywhere
in versioning code or otherwise, so it should be safely removable.

Reduces the size of this struct by 8 bytes.

Pull Request: https://projects.blender.org/blender/blender/pulls/143874