Test: LibOverride: Add more complex case for multi-level hierarchy handling.
Add some initial data to test proper handling of resync when there are
specific changes in the reference collection hierarchy (in particular,
when the last child of a given collection is moved somewhere else).

Also will check for correct handling of recursive liboverrides resync in
that case.

The current failing part is commented out for now, until fix is
committed.