Fix #144364: Incorrect RNA update function called for Bones
The issue was that an rna function was called on `Bone`
that expected an `EditBone`. Because the pointer was just casted,
this means it would write into wrong bytes on the struct.

Caused by a43359eb88e898c8e1960564705bf80104b60295

Pull Request: https://projects.blender.org/blender/blender/pulls/144447