Fix #144542: Workbench DoF crash when using Vulkan
The depth of field shader create info contained a resource that wasn't
used in the shader or fulfilled. In the Vulkan backend the shader resource
list is used in the outer loop to find out which resources needs to be
requested. As this resource didn't exist it asserted in the Vulkan backend
for incorrect usage.

Pull Request: https://projects.blender.org/blender/blender/pulls/144606