Cleanup: Mesh: Use edge_other_vert instead of ternary operator
Its generally preferable to use `edge_other_vert` if you have a set of
2 ints and know some of them and want to select other. This is
preferable since it will make a check that assumption is met and at
least one of the ints is the same as the known one. This function itself
is also made to work with topological indices so has invariant for
positive ints and can work without branches and even test instructions.

Pull Request: https://projects.blender.org/blender/blender/pulls/144631